refactor(bigtable): untemplate `AsyncRowReader`
